+++ src/javascript/constructors/editorclient.js
@@ -1,0 +1,15 @@
+
+
+function EditorClient({
+    name="",
+},object){
+    return {
+        do(name,args){
+            if(object[name]!=null){
+                object[name](null,args)
+            }
+        }
+    }
+}
+
+export default EditorClient