+++ src/javascript/constructors/tab.js
@@ -1,0 +1,14 @@
+function Tab({
+    title,
+    isEditor,
+    path,
+    component
+}){
+    if(isEditor){
+        
+    }else{
+
+    }
+}
+
+export default Tab