+++ src/javascript/utils/static.config.js
@@ -1,5 +1,4 @@
 import {puffin} from '@mkenzo_8/puffin'
-import ThemeProvider from './themeprovider'
 import {initConfiguration , getConfiguration} from './configurator'
 
 initConfiguration()
@@ -7,10 +6,10 @@ const cachedConfiguration = getConfiguration()
 
 let editorConf = Object.assign({},cachedConfiguration.editor)
 
-const ConfigProvider = puffin.state(editorConf)
+const StaticConfig = puffin.state(editorConf)
 
-ConfigProvider.changed(function(data){
+StaticConfig.changed(function(data){
     cachedConfiguration.store.set('editor',data)
 })
 
-export default ConfigProvider
+export default StaticConfig